fixed medical history flow from service to view model
