+++ api/serializers/medication_history_serializer.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+class MedicationHistorySerializer(serializers.Serializer):
+    approval_name = serializers.CharField()
+    doctor_name = serializers.CharField()
+    patient_name = serializers.CharField()
+    qty_changed = serializers.IntegerField()
+    qty_remaining = serializers.IntegerField()
+    date = serializers.DateTimeField()
